+++ changelog.md
@@ -1,3 +1,7 @@
+- ???
+
+    Added `nubOrd` and friends.
+
 - 0.2.1.0
 
     Adding `Semigroup` instances for GHC 8.4.